requirements.txt
Pillow==9.1.0
pynput==1.7.3
pyinstaller==5.6.2
tkinter==8.6.9